+++ app/Main.hs
@@ -3,6 +3,7 @@
 
 module Main where
 
+import Parsing
 import Path
 import qualified Turtle hiding (switch)
 import Prelude hiding (putStrLn)
@@ -20,11 +21,12 @@ import Control.Monad (when)
 import qualified Data.Text.IO as T
 
 data ImportParams = ImportParams { maybeImportBaseDir :: Maybe TurtlePath
-                                 , importUseRunDir :: Bool
+                                 , importStartYear :: Maybe String
                                  , onlyNewFiles :: Bool
+                                 , importUseRunDir :: Bool
                                  } deriving (Show)
 
-data ReportParams = ReportParams { maybeReportBaseDir :: Maybe TurtlePath } deriving (Show)
+newtype ReportParams = ReportParams {maybeReportBaseDir :: Maybe TurtlePath} deriving Show
 
 data Command = Import ImportParams | Report ReportParams deriving (Show)
 
@@ -45,6 +47,7 @@ main = do
 
 toRuntimeOptionsImport :: MainParams -> ImportParams -> IO RT.RuntimeOptions
 toRuntimeOptionsImport mainParams' subParams' = do
+  startYear <- parseStartYear $ importStartYear subParams'
   let maybeBD = maybeImportBaseDir subParams' :: Maybe TurtlePath
   Control.Monad.when (importUseRunDir subParams') $ do
     T.putStrLn "The enable-future-rundir option is now the default, no need to specify it. This option is currently being ignored and will be removed in future."
@@ -52,6 +55,7 @@ toRuntimeOptionsImport mainParams' subParams' = do
   hli <- hledgerInfoFromPath $ hledgerPathOpt mainParams'
   return RT.RuntimeOptions { RT.baseDir = bd
                            , RT.importRunDir = runDir
+                           , RT.importStartYear = startYear
                            , RT.onlyNewFiles = onlyNewFiles subParams'
                            , RT.hfVersion = versionInfo'
                            , RT.hledgerInfo = hli
@@ -67,6 +71,7 @@ toRuntimeOptionsReport mainParams' subParams' = do
   hli <- hledgerInfoFromPath $ hledgerPathOpt mainParams'
   return RT.RuntimeOptions { RT.baseDir = bd
                            , RT.importRunDir = [reldir|.|]
+                           , RT.importStartYear = Nothing
                            , RT.onlyNewFiles = False
                            , RT.hfVersion = versionInfo'
                            , RT.hledgerInfo = hli
@@ -93,8 +98,9 @@ verboseParser = MainParams
 subcommandParserImport :: Parser ImportParams
 subcommandParserImport = ImportParams
   <$> optional (Turtle.argPath "dir" "The directory to import. Use the base directory for a full import or a sub-directory for a partial import. Defaults to the current directory. This behaviour is changing: see --enable-future-rundir")
-  <*> switch (long "enable-future-rundir" <> help "This switch is currently being ignored, since the behaviour it previously enabled is now the default. It will be removed in future.")
+  <*> optional (strOption (long "start-year" <> metavar "YEAR" <> help "Import only from the specified year and onwards, ignoring previous years. By default all available years are imported. Valid values include a 4-digit year or 'current' for the current year"))
   <*> switch (long "new-files-only" <> help "Don't regenerate transaction files if they are already present. This applies to hledger journal files as well as files produced by the preprocess and construct scripts.")
+  <*> switch (long "enable-future-rundir" <> help "This switch is currently being ignored, since the behaviour it previously enabled is now the default. It will be removed in future.")
 
 subcommandParserReport :: Parser ReportParams
 subcommandParserReport = ReportParams